requirements_full.txt
# Trading Bot Requirements
# Core Framework
flask==2.3.3
python-telegram-bot==20.6
fastapi==0.104.1
uvicorn==0.24.0

# Solana & Blockchain
solana==0.30.2
solders==0.18.1
anchorpy==0.19.1

# HTTP & Async
aiohttp==3.9.1
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Encoding/Decoding
base64==0.1.0
base58==2.1.1

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging & Monitoring
structlog==23.2.0

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development Tools (Optional)
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Discord & Social Integrations (from original)
discord-webhook==1.3.0
slack-webhook==1.0.6
tweepy==4.14.0

# Email Support (from original)
smtplib2==0.2.1

# Scheduling
APScheduler==3.10.4

# Database (Optional - for storing user data)
sqlalchemy==2.0.23
alembic==1.13.1
sqlite==0.1.0  # For local development

# Monitoring & Performance
psutil==5.9.6
memory-profiler==0.61.0

# Security
cryptography==41.0.8
bcrypt==4.1.2

# Web Server
waitress==2.1.2
gunicorn==21.2.0

# Jupyter (for analysis - optional)
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0

# API Rate Limiting
ratelimit==2.2.1
slowapi==0.1.9

# JSON & Data Validation
jsonschema==4.20.0
marshmallow==3.20.2

# Time & Date Utilities
python-dateutil==2.8.2
pytz==2023.3